Housekeeping: separated test steps into different files.
