+++ tests/features/steps/wrappers.py
@@ -1,0 +1,13 @@
+# -*- coding: utf-8
+import re
+
+
+def expect_exact(context, expected, timeout):
+    try:
+        context.cli.expect_exact(expected, timeout=timeout)
+    except:
+        # Strip color codes out of the output.
+        actual = re.sub(r'\x1b\[([0-9A-Za-z;?])+[m|K]?', '', context.cli.before)
+        raise Exception('Expected:\n---\n{0!r}\n---\n\nActual:\n---\n{1!r}\n---'.format(
+            expected,
+            actual))
